// frontend/src/types/AttributeType.ts
export enum AttributeType {
    ATTRIBUTE = 'ATTRIBUTE',
    CASE_ID = 'CASE_ID',
    ACTIVITY = 'ACTIVITY',
    RESOURCE = 'RESOURCE',
    ROLE = 'ROLE',
    START_TIMESTAMP = 'START_TIMESTAMP',
    COMPLETE_TIMESTAMP = 'COMPLETE_TIMESTAMP',
    CASE_ATTRIBUTE = 'CASE_ATTRIBUTE',
    EVENT_ATTRIBUTE = 'EVENT_ATTRIBUTE',
    IGNORE_ATTRIBUTE = 'IGNORE_ATTRIBUTE',
}